ai-cfo-backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
python-dotenv
gunicorn
whitenoise
azure-identity
azure-keyvault-secrets
prophet
scikit-learn
pandas
numpy
google-generativeai
